Move some enums and types out of types package
