+++ packages/server/src/attestation/verifications/verifyPacked.ts
@@ -2,14 +2,14 @@ import base64url from 'base64url';
 import cbor from 'cbor';
 import elliptic from 'elliptic';
 import NodeRSA, { SigningSchemeHash } from 'node-rsa';
-import {
-  AttestationObject,
-  VerifiedAttestation,
-  COSEKEYS,
-  COSEPublicKey as COSEPublicKeyType,
-} from '@simplewebauthn/typescript-types';
 
-import convertCOSEtoPKCS from '../../helpers/convertCOSEtoPKCS';
+import type { AttestationObject } from '../../helpers/decodeAttestationObject';
+import type { VerifiedAttestation } from '../verifyAttestationResponse';
+
+import convertCOSEtoPKCS, {
+  COSEKEYS,
+  COSEPublicKey as COSEPublicKeyType
+} from '../../helpers/convertCOSEtoPKCS';
 import toHash from '../../helpers/toHash';
 import convertASN1toPEM from '../../helpers/convertASN1toPEM';
 import getCertificateInfo from '../../helpers/getCertificateInfo';
