+++ packages/server/src/helpers/decodeAttestationObject.ts
@@ -1,6 +1,5 @@
 import base64url from 'base64url';
 import cbor from 'cbor';
-import { AttestationObject } from '@simplewebauthn/typescript-types';
 
 /**
  * Convert an AttestationObject from base64 string to a proper object
@@ -14,3 +13,20 @@ export default function decodeAttestationObject(
   const toCBOR: AttestationObject = cbor.decodeAllSync(toBuffer)[0];
   return toCBOR;
 }
+
+export enum ATTESTATION_FORMATS {
+  FIDO_U2F = 'fido-u2f',
+  PACKED = 'packed',
+  ANDROID_SAFETYNET = 'android-safetynet',
+  NONE = 'none',
+}
+
+export type AttestationObject = {
+  fmt: ATTESTATION_FORMATS;
+  attStmt: {
+    sig?: Buffer;
+    x5c?: Buffer[];
+    response?: Buffer;
+  };
+  authData: Buffer;
+};
